fixing more warnings, explicit casts, etc
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@4874 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ common/include/pcl/common/time.h
@@ -143,7 +143,7 @@ getTime ()
 {
   boost::posix_time::ptime epoch_time (boost::gregorian::date (1970, 1, 1));
   boost::posix_time::ptime current_time = boost::posix_time::microsec_clock::local_time ();
-  return (current_time - epoch_time).total_nanoseconds () * 1.0e-9;
+  return (static_cast<double>((current_time - epoch_time).total_nanoseconds ()) * 1.0e-9);
 }
 
 /// Executes code, only if secs are gone since last exec.
